requirements.txt
python-dotenv==1.0.0
google-generativeai==0.8.3
flask==3.0.0
